# zoekt: handle partially indexed RepositoryRevisions