This commit fixes `IndexedRepoRevs.add` to correctly filter out unindexed `search.RevisionSpecifiers` from a `*search.RepositoryRevisions` with *some* indexed rev specs.

Benchmarks show no real diff:

```
name              old time/op    new time/op    delta
SearchResults-12     1.33s ± 7%     1.29s ± 2%   ~     (p=0.114 n=9+7)

name              old alloc/op   new alloc/op   delta
SearchResults-12     469MB ± 0%     469MB ± 0%   ~     (p=0.955 n=7+8)

name              old allocs/op  new allocs/op  delta
SearchResults-12     5.52M ± 0%     5.52M ± 0%   ~     (p=0.650 n=9+9)
```


<!-- Reminder: Have you updated the changelog and relevant docs (user docs, architecture diagram, etc) ? -->
<!-- Please notify @distribution if this PR contains changes to CI that may need to be cherry-picked on to patch release branches -->